Изменение структуры рейтинга и типов возвращаемых значений
Удалены классы `RatingResponse` и `RatingItem` из `RatingResponse.cs`.
Добавлено свойство `Products` в `RatingItem`, заменяющее `Result` и `Groups`.
Изменен тип возвращаемого значения метода `GetContentRatingAsync` в интерфейсе `IOzonApiService` с `Task<int>` на `Task<double>`.
Обновлена логика обработки ответа в классе `OzonApiService` для работы с `Products` вместо `Result`.

+++ Services/Ozon/IOzonApiService.cs
@@ -7,7 +7,7 @@ namespace ai_it_wiki.Services.Ozon
 {
   public interface IOzonApiService
   {
-    Task<int> GetContentRatingAsync(string sku, CancellationToken cancellationToken = default);
+    Task<double> GetContentRatingAsync(string sku, CancellationToken cancellationToken = default);
     Task<string> GetProductInfoAsync(string sku, CancellationToken cancellationToken = default);
     Task<string> GetProductDescriptionAsync(string sku, CancellationToken cancellationToken = default);
     Task<string> ImportProductAsync(string sku, string improvedContent, CancellationToken cancellationToken = default);
